more refactoring to new pattern
