--- warp-drive-packages/core/src/reactive/-private/fields/compute.ts
@@ -1,285 +1,0 @@
-import { DEBUG } from '@warp-drive/core/build-config/env';
-
-import type { Store, StoreRequestInput } from '../../../index.ts';
-import type { Future } from '../../../request.ts';
-import { defineSignal, RelatedCollection as ManyArray } from '../../../store/-private.ts';
-import { getOrSetGlobal } from '../../../types/-private.ts';
-import type { Cache } from '../../../types/cache.ts';
-import type { ResourceRelationship as SingleResourceRelationship } from '../../../types/cache/relationship.ts';
-import type { StableRecordIdentifier } from '../../../types/identifier.ts';
-import type { ObjectValue } from '../../../types/json/raw.ts';
-import type { FieldSchema, LegacyHasManyField, ObjectField, SchemaObjectField } from '../../../types/schema/fields.ts';
-import type { CollectionResourceRelationship, Link, Links } from '../../../types/spec/json-api-raw.ts';
-import { RecordStore } from '../../../types/symbols.ts';
-import { ReactiveResource } from '../record.ts';
-import type { SchemaService } from '../schema.ts';
-import { Editable, Identifier, Legacy, Parent } from '../symbols.ts';
-import { getFieldCacheKeyStrict } from './get-field-key.ts';
-import type { ManagedArray } from './managed-array.ts';
-import { ManagedObject } from './managed-object.ts';
-import { ManyArrayManager } from './many-array-manager.ts';
-
-export const ManagedArrayMap: Map<ReactiveResource, Map<string, ManagedArray | ManyArray>> = getOrSetGlobal(
-  'ManagedArrayMap',
-  new Map<ReactiveResource, Map<string, ManagedArray | ManyArray>>()
-);
-export const ManagedObjectMap: Map<ReactiveResource, Map<string, ManagedObject | ReactiveResource>> = getOrSetGlobal(
-  'ManagedObjectMap',
-  new Map<ReactiveResource, Map<string, ManagedObject | ReactiveResource>>()
-);
-
-export function peekManagedArray(record: ReactiveResource, field: FieldSchema): ManyArray | ManagedArray | undefined {
-  const managedArrayMapForRecord = ManagedArrayMap.get(record);
-  if (managedArrayMapForRecord) {
-    return managedArrayMapForRecord.get(field.name);
-  }
-}
-
-export function peekManagedObject(record: ReactiveResource, field: ObjectField): ManagedObject | undefined;
-export function peekManagedObject(record: ReactiveResource, field: SchemaObjectField): ReactiveResource | undefined;
-export function peekManagedObject(
-  record: ReactiveResource,
-  field: ObjectField | SchemaObjectField
-): ManagedObject | ReactiveResource | undefined {
-  const managedObjectMapForRecord = ManagedObjectMap.get(record);
-  if (managedObjectMapForRecord) {
-    return managedObjectMapForRecord.get(field.name);
-  }
-}
-
-export function computeObject(
-  schema: SchemaService,
-  cache: Cache,
-  record: ReactiveResource,
-  identifier: StableRecordIdentifier,
-  field: ObjectField,
-  path: string[],
-  editable: boolean,
-  legacy: boolean
-): ManagedObject | null {
-  const managedObjectMapForRecord = ManagedObjectMap.get(record);
-  let managedObject;
-  if (managedObjectMapForRecord) {
-    managedObject = managedObjectMapForRecord.get(field.name);
-  }
-  if (managedObject) {
-    return managedObject as ManagedObject;
-  } else {
-    let rawValue = (editable ? cache.getAttr(identifier, path) : cache.getRemoteAttr(identifier, path)) as object;
-    if (!rawValue) {
-      return null;
-    }
-    if (field.type) {
-      const transform = schema.transformation(field);
-      rawValue = transform.hydrate(rawValue as ObjectValue, field.options ?? null, record) as object;
-    }
-    managedObject = new ManagedObject(schema, cache, field, rawValue, identifier, path, record, editable, legacy);
-
-    if (!managedObjectMapForRecord) {
-      ManagedObjectMap.set(record, new Map([[field.name, managedObject]]));
-    } else {
-      managedObjectMapForRecord.set(field.name, managedObject);
-    }
-  }
-  return managedObject;
-}
-
-export function computeSchemaObject(
-  store: Store,
-  cache: Cache,
-  record: ReactiveResource,
-  identifier: StableRecordIdentifier,
-  field: SchemaObjectField,
-  path: string[],
-  legacy: boolean,
-  editable: boolean
-): ReactiveResource | null {
-  const schemaObjectMapForRecord = ManagedObjectMap.get(record);
-  let schemaObject;
-  if (schemaObjectMapForRecord) {
-    schemaObject = schemaObjectMapForRecord.get(field.name);
-  }
-  if (schemaObject) {
-    return schemaObject as ReactiveResource;
-  } else {
-    const rawValue = (editable ? cache.getAttr(identifier, path) : cache.getRemoteAttr(identifier, path)) as object;
-    if (!rawValue) {
-      return null;
-    }
-    const embeddedPath = path.slice();
-    schemaObject = new ReactiveResource(
-      store,
-      identifier,
-      {
-        [Editable]: editable,
-        [Legacy]: legacy,
-      },
-      true,
-      field,
-      embeddedPath
-    );
-  }
-  if (!schemaObjectMapForRecord) {
-    ManagedObjectMap.set(record, new Map([[field.name, schemaObject]]));
-  } else {
-    schemaObjectMapForRecord.set(field.name, schemaObject);
-  }
-  return schemaObject;
-}
-
-interface ResourceRelationship<T extends ReactiveResource = ReactiveResource> {
-  lid: string;
-  [Parent]: ReactiveResource;
-  [RecordStore]: Store;
-  name: string;
-
-  data: T | null;
-  links: Links;
-  meta: Record<string, unknown>;
-}
-
-// TODO probably this should just be a Document
-// but its separate until we work out the lid situation
-class ResourceRelationship<T extends ReactiveResource = ReactiveResource> {
-  constructor(
-    store: Store,
-    cache: Cache,
-    parent: ReactiveResource,
-    identifier: StableRecordIdentifier,
-    field: FieldSchema,
-    name: string,
-    editable: boolean
-  ) {
-    const rawValue = (
-      editable ? cache.getRelationship(identifier, name) : cache.getRemoteRelationship(identifier, name)
-    ) as SingleResourceRelationship;
-
-    // TODO setup true lids for relationship documents
-    // @ts-expect-error we need to give relationship documents a lid
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    this.lid = rawValue.lid ?? rawValue.links?.self ?? `relationship:${identifier.lid}.${name}`;
-    this.data = rawValue.data ? store.peekRecord<T>(rawValue.data) : null;
-    this.name = name;
-
-    if (DEBUG) {
-      this.links = Object.freeze(Object.assign({}, rawValue.links));
-      this.meta = Object.freeze(Object.assign({}, rawValue.meta));
-    } else {
-      this.links = rawValue.links ?? {};
-      this.meta = rawValue.meta ?? {};
-    }
-
-    this[RecordStore] = store;
-    this[Parent] = parent;
-  }
-
-  fetch(options?: StoreRequestInput<T, T>): Future<T> {
-    const url = options?.url ?? getHref(this.links.related) ?? getHref(this.links.self) ?? null;
-
-    if (!url) {
-      throw new Error(
-        `Cannot ${options?.method ?? 'fetch'} ${this[Parent][Identifier].type}.${String(
-          this.name
-        )} because it has no related link`
-      );
-    }
-    const request = Object.assign(
-      {
-        url,
-        method: 'GET',
-      },
-      options
-    );
-
-    return this[RecordStore].request<T>(request);
-  }
-}
-
-defineSignal(ResourceRelationship.prototype, 'data', null);
-defineSignal(ResourceRelationship.prototype, 'links', null);
-defineSignal(ResourceRelationship.prototype, 'meta', null);
-
-function getHref(link?: Link | null): string | null {
-  if (!link) {
-    return null;
-  }
-  if (typeof link === 'string') {
-    return link;
-  }
-  return link.href;
-}
-
-export function computeResource<T extends ReactiveResource>(
-  store: Store,
-  cache: Cache,
-  parent: ReactiveResource,
-  identifier: StableRecordIdentifier,
-  field: FieldSchema,
-  prop: string,
-  editable: boolean
-): ResourceRelationship<T> {
-  if (field.kind !== 'resource') {
-    throw new Error(`The schema for ${identifier.type}.${String(prop)} is not a resource relationship`);
-  }
-
-  return new ResourceRelationship<T>(store, cache, parent, identifier, field, prop, editable);
-}
-
-export function computeHasMany(
-  store: Store,
-  schema: SchemaService,
-  cache: Cache,
-  record: ReactiveResource,
-  identifier: StableRecordIdentifier,
-  field: LegacyHasManyField,
-  path: string[],
-  editable: boolean,
-  legacy: boolean
-): ManyArray | null {
-  // the thing we hand out needs to know its owner and path in a private manner
-  // its "address" is the parent identifier (identifier) + field name (field.name)
-  //  in the nested object case field name here is the full dot path from root resource to this value
-  // its "key" is the field on the parent record
-  // its "owner" is the parent record
-
-  const managedArrayMapForRecord = ManagedArrayMap.get(record);
-  let managedArray: ManyArray | undefined;
-  if (managedArrayMapForRecord) {
-    managedArray = managedArrayMapForRecord.get(getFieldCacheKeyStrict(field)) as ManyArray | undefined;
-  }
-  if (managedArray) {
-    return managedArray;
-  } else {
-    const rawValue = cache.getRelationship(identifier, getFieldCacheKeyStrict(field)) as CollectionResourceRelationship;
-    if (!rawValue) {
-      return null;
-    }
-    managedArray = new ManyArray<unknown>({
-      store,
-      type: field.type,
-      identifier,
-      cache,
-      field: legacy ? field : undefined,
-      // we divorce the reference here because ManyArray mutates the target directly
-      // before sending the mutation op to the cache. We may be able to avoid this in the future
-      identifiers: rawValue.data?.slice() as StableRecordIdentifier[],
-      key: field.name,
-      meta: rawValue.meta || null,
-      links: rawValue.links || null,
-      isPolymorphic: field.options.polymorphic ?? false,
-      isAsync: field.options.async ?? false,
-      // TODO: Grab the proper value
-      _inverseIsAsync: false,
-      // @ts-expect-error Typescript doesn't have a way for us to thread the generic backwards so it infers unknown instead of T
-      manager: new ManyArrayManager(record, editable),
-      isLoaded: true,
-      allowMutation: editable,
-    });
-    if (!managedArrayMapForRecord) {
-      ManagedArrayMap.set(record, new Map([[field.name, managedArray]]));
-    } else {
-      managedArrayMapForRecord.set(field.name, managedArray);
-    }
-  }
-  return managedArray;
-}
