--- warp-drive-packages/core/src/reactive/-private/kind/-utils.ts
@@ -1,1 +1,0 @@
-export type PathLike = string | symbol | Array<string | symbol>;
