+++ warp-drive-packages/core/src/reactive/-private/kind/belongs-to-field.ts
@@ -1,6 +1,12 @@
+import { assert } from '@warp-drive/build-config/macros';
+
 import type { Store } from '../../../store/-private';
 import type { StableRecordIdentifier } from '../../../types';
 import type { LegacyBelongsToField } from '../../../types/schema/fields';
+import type { SingleResourceRelationship } from '../../../types/spec/json-api-raw';
+import type { ModeInfo } from '../default-mode';
+import { getFieldCacheKeyStrict } from '../fields/get-field-key';
+import type { SchemaService } from '../schema';
 
 export function getBelongsToField(
   store: Store,
@@ -8,10 +14,20 @@ export function getBelongsToField(
   resourceKey: StableRecordIdentifier,
   field: LegacyBelongsToField,
   path: string | string[],
-  editable: boolean
+  mode: ModeInfo
 ): unknown {
-  const { cache } = store;
-  return editable ? cache.getAttr(resourceKey, path) : cache.getRemoteAttr(resourceKey, path);
+  const { schema, cache } = store;
+  if (field.options.linksMode) {
+    const rawValue = mode.editable
+      ? (cache.getRelationship(resourceKey, getFieldCacheKeyStrict(field)) as SingleResourceRelationship)
+      : (cache.getRemoteRelationship(resourceKey, getFieldCacheKeyStrict(field)) as SingleResourceRelationship);
+
+    return rawValue.data ? store.peekRecord(rawValue.data) : null;
+  }
+
+  // FIXME move this to a "LegacyMode" make this part of "PolarisMode"
+  assert(`Can only use belongsTo fields when the resource is in legacy mode`, mode.legacy);
+  return (schema as SchemaService)._kind('@legacy', 'belongsTo').get(store, record, resourceKey, field);
 }
 
 export function setBelongsToField(
@@ -20,9 +36,12 @@ export function setBelongsToField(
   resourceKey: StableRecordIdentifier,
   field: LegacyBelongsToField,
   path: string | string[],
+  mode: ModeInfo,
   value: unknown
 ): boolean {
-  const { cache } = store;
+  const { schema } = store;
 
+  assert(`Can only mutate belongsTo fields when the resource is in legacy mode`, mode.legacy);
+  (schema as SchemaService)._kind('@legacy', 'belongsTo').set(store, record, resourceKey, field, value);
   return true;
 }
