+++ warp-drive-packages/core/src/reactive/-private/kind/local-field.ts
@@ -7,7 +7,7 @@ import {
 } from '../../../store/-private';
 import type { StableRecordIdentifier } from '../../../types';
 import type { LocalField } from '../../../types/schema/fields';
-import type { PathLike } from './-utils';
+import type { ModeInfo, PathLike } from '../default-mode';
 
 export function getLocalField(
   store: Store,
@@ -15,7 +15,7 @@ export function getLocalField(
   resourceKey: StableRecordIdentifier,
   field: LocalField,
   path: PathLike,
-  editable: boolean
+  mode: ModeInfo
 ): unknown {
   const signals = withSignalStore(record);
   const prop = Array.isArray(path) ? path.at(-1)! : path;
@@ -30,6 +30,7 @@ export function setLocalField(
   resourceKey: StableRecordIdentifier,
   field: LocalField,
   path: PathLike,
+  mode: ModeInfo,
   value: unknown
 ): boolean {
   const signals = withSignalStore(record);
